maloney-plugins/src/main/java/ch/hsr/maloney/maloney_plugins/AuthenticodePEJob.java
package ch.hsr.maloney.maloney_plugins;

import ch.hsr.maloney.processing.Job;
import ch.hsr.maloney.processing.JobCancelledException;
import ch.hsr.maloney.util.Context;
import ch.hsr.maloney.util.Event;
import org.apache.logging.log4j.Logger;

import java.io.FileInputStream;
import java.io.IOException;
import java.io.InputStream;
import java.util.LinkedList;
import java.util.List;

/**
 * Implementation which examines Authenticode information of a portable executable.
 */
public class AuthenticodePEJob implements Job {
    private static final String JOB_NAME = "AuthenticodePEJob";
    private static final String NEW_FILE_EVENT_NAME = "newFile";
    private final LinkedList<String> producedEvents;
    private final LinkedList<String> requiredEvents;
    private final Logger logger;

    public AuthenticodePEJob() {
        producedEvents = new LinkedList<>();
        requiredEvents = new LinkedList<>();
        requiredEvents.add(NEW_FILE_EVENT_NAME);
        logger = org.apache.logging.log4j.LogManager.getLogger();
    }

    @Override
    public boolean shouldRun(Context ctx, Event evt) {
        try {
            InputStream is = new FileInputStream(ctx.getDataSource().getFile(evt.getFileUuid()));

            // First two bytes are 4D 5A or 77 90 as integers, also known as MZ
            byte[] buffer = new byte[2];
            int bytesRead = is.read(buffer);
            is.close();
            return bytesRead == 2
                    && buffer[0] == 77
                    && buffer[1] == 90;
        } catch (IOException e) {
            logger.warn("Could not identify portable executable.", e);
        }
        return false;
    }

    @Override
    public boolean canRun(Context ctx, Event evt) {
        return true;
    }

    @Override
    public List<Event> run(Context ctx, Event evt) throws JobCancelledException {
        // TODO implement
        return null;
    }

    @Override
    public List<String> getRequiredEvents() {
        return requiredEvents;
    }

    @Override
    public List<String> getProducedEvents() {
        return producedEvents;
    }

    @Override
    public String getJobName() {
        return JOB_NAME;
    }

    @Override
    public String getJobConfig() {
        return null;
    }

    @Override
    public void setJobConfig(String config) {

    }
}
